CLK32-889:Add SocialLogin
