+++ ClearKeep/Scenes/ContentView.swift
@@ -23,7 +23,7 @@ struct ContentView: View {
 			if isRunningTests {
 				Text("Running unit tests")
 			} else {
-				HomeView()
+				SocialConfirm()
 					.inject(container)
 			}
 		}
